+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### New Features
 * [#182][gh-182] Support file name aliases via `#+ROAM_ALIAS`.
+* [#188][gh-188] Add `org-roam-protocol`, shifting `roam://` link handling into Emacs-lisp.
 
 ### Features
 * [#165][gh-165] Add templating functionality via `org-roam-templates`.
@@ -97,6 +98,7 @@ Mostly a documentation/cleanup release.
 [gh-143]: https://github.com/jethrokuan/org-roam/pull/143
 [gh-165]: https://github.com/jethrokuan/org-roam/pull/165
 [gh-182]: https://github.com/jethrokuan/org-roam/pull/182
+[gh-188]: https://github.com/jethrokuan/org-roam/pull/188
 
  # Local Variables:
  # eval: (auto-fill-mode -1)